Added Random Quest and Button Class
Do you ever think "Oh I should do that soon" or "I keep putting this off".  "If only there was a way to remember things I want to do or hold myself accountable."  Introducing Random Quests!  An easy way to get to the tasks that keep eluding you.  I intend on the user being able to insert and delete Random Quests themselves in future. I also made a long overdue button class that can take the positions of a rectangle and generate a button press, maybe in future it can draw the rectangle and work for other shapes too.